+++ src/oinfo.cpp
@@ -238,7 +238,7 @@ void Info::init_random_seed(int randomSeed)
 		random_seed = randomSeed;
 	else
 	{
-		randomSeed = time(NULL);
+		randomSeed = (int) time(NULL);
 		randomSeed = (int) _rotr( randomSeed, 4 );
 		if( randomSeed < 0 )
 			randomSeed = ~randomSeed;
